chore: refining new APIResourceCatalog (focus on removing dead or almost dead code paths)
