--- internal/watch/discovery_integration_test.go
@@ -1,357 +1,0 @@
-/*
-SPDX-License-Identifier: Apache-2.0
-
-Copyright 2025 ConfigButler
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package watch
-
-import (
-	"context"
-	"testing"
-	"time"
-
-	"github.com/go-logr/logr"
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-
-	configv1alpha1 "github.com/ConfigButler/gitops-reverser/api/v1alpha1"
-	"github.com/ConfigButler/gitops-reverser/internal/rulestore"
-)
-
-// TestCRDDiscoveryLifecycle tests that the watch manager discovers newly installed CRDs
-// when ReconcileForRuleChange is called after CRD installation.
-func TestCRDDiscoveryLifecycle(t *testing.T) {
-	// Setup scheme with CRD types
-	scheme := runtime.NewScheme()
-	_ = clientgoscheme.AddToScheme(scheme)
-	_ = configv1alpha1.AddToScheme(scheme)
-	_ = apiextensionsv1.AddToScheme(scheme)
-
-	ctx := context.Background()
-
-	// Create fake client
-	fakeClient := fake.NewClientBuilder().
-		WithScheme(scheme).
-		Build()
-
-	// Create rule store and watch manager.
-	//
-	// discoveryFilter is stubbed so the test is hermetic. The real
-	// FilterDiscoverableGVRs reaches whatever cluster is in the ambient
-	// kubeconfig via ctrl.GetConfig(); a fake client has no discovery endpoint,
-	// so without this stub the result would depend on the developer's
-	// environment (e.g. a running e2e cluster that happens to have the CRD).
-	// Returning nil models "nothing discoverable" — actual discovery behaviour
-	// is covered by the e2e suite.
-	ruleStore := rulestore.NewStore()
-	manager := &Manager{
-		Client:          fakeClient,
-		Log:             logr.Discard(),
-		RuleStore:       ruleStore,
-		discoveryFilter: func(context.Context, []GVR) []GVR { return nil },
-		resourceCatalog: newCommonTestCatalog(t),
-		discoveryClient: commonTestDiscoveryClient(),
-	}
-
-	// Step 1: Create a WatchRule that references a CRD resource that doesn't exist yet
-	watchRule := &configv1alpha1.WatchRule{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-icecream-rule",
-			Namespace: "default",
-		},
-		Spec: configv1alpha1.WatchRuleSpec{
-			TargetRef: configv1alpha1.LocalTargetReference{
-				Name: "test-dest",
-			},
-			Rules: []configv1alpha1.ResourceRule{
-				{
-					APIGroups:   []string{"shop.example.com"},
-					APIVersions: []string{"v1"},
-					Resources:   []string{"icecreamorders"},
-					Operations:  []configv1alpha1.OperationType{configv1alpha1.OperationUpdate},
-				},
-			},
-		},
-	}
-
-	// Add rule to store (simulating what the WatchRule controller does)
-	ruleStore.AddOrUpdateWatchRule(
-		*watchRule,
-		"test-dest", "default",
-		"test-repo", "default",
-		"main",
-		"test-folder",
-	)
-
-	// Step 2: Call ReconcileForRuleChange - CRD doesn't exist yet
-	t.Log("Step 1: Reconcile before CRD exists")
-	if err := manager.ReconcileForRuleChange(ctx); err != nil {
-		t.Fatalf("Initial reconcile failed: %v", err)
-	}
-
-	// Verify no informers started (CRD not discoverable)
-	manager.informersMu.Lock()
-	initialInformerCount := len(manager.activeInformers)
-	manager.informersMu.Unlock()
-
-	if initialInformerCount != 0 {
-		t.Errorf("Expected 0 active informers before CRD exists, got %d", initialInformerCount)
-	}
-
-	// Step 3: Install the CRD (simulate kubectl apply -f crd.yaml)
-	t.Log("Step 2: Installing IceCreamOrder CRD")
-	crd := &apiextensionsv1.CustomResourceDefinition{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "icecreamorders.shop.example.com",
-		},
-		Spec: apiextensionsv1.CustomResourceDefinitionSpec{
-			Group: "shop.example.com",
-			Names: apiextensionsv1.CustomResourceDefinitionNames{
-				Plural:   "icecreamorders",
-				Singular: "icecreamorder",
-				Kind:     "IceCreamOrder",
-			},
-			Scope: apiextensionsv1.NamespaceScoped,
-			Versions: []apiextensionsv1.CustomResourceDefinitionVersion{
-				{
-					Name:    "v1",
-					Served:  true,
-					Storage: true,
-					Schema: &apiextensionsv1.CustomResourceValidation{
-						OpenAPIV3Schema: &apiextensionsv1.JSONSchemaProps{
-							Type: "object",
-							Properties: map[string]apiextensionsv1.JSONSchemaProps{
-								"spec": {
-									Type: "object",
-									Properties: map[string]apiextensionsv1.JSONSchemaProps{
-										"customerName": {Type: "string"},
-										"flavor":       {Type: "string"},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-		Status: apiextensionsv1.CustomResourceDefinitionStatus{
-			Conditions: []apiextensionsv1.CustomResourceDefinitionCondition{
-				{
-					Type:   apiextensionsv1.Established,
-					Status: apiextensionsv1.ConditionTrue,
-				},
-			},
-		},
-	}
-
-	if err := fakeClient.Create(ctx, crd); err != nil {
-		t.Fatalf("Failed to create CRD: %v", err)
-	}
-
-	// Step 4: Trigger immediate reconciliation (simulating what the fix will do)
-	t.Log("Step 3: Reconcile after CRD exists")
-	if err := manager.ReconcileForRuleChange(ctx); err != nil {
-		t.Fatalf("Post-CRD reconcile failed: %v", err)
-	}
-
-	// Note: In a real cluster with actual API discovery, the CRD would now be discoverable.
-	// The fake client doesn't support full API discovery, so we're testing the flow.
-	// The e2e tests will verify actual discovery behavior.
-
-	t.Log("✅ CRD discovery lifecycle test completed - verifies reconciliation flow")
-}
-
-// TestUnavailableGVRTracking tests that the manager tracks GVRs that are requested but not discoverable.
-func TestUnavailableGVRTracking(t *testing.T) {
-	scheme := runtime.NewScheme()
-	_ = clientgoscheme.AddToScheme(scheme)
-	_ = configv1alpha1.AddToScheme(scheme)
-
-	ctx := context.Background()
-
-	fakeClient := fake.NewClientBuilder().
-		WithScheme(scheme).
-		Build()
-
-	ruleStore := rulestore.NewStore()
-	manager := &Manager{
-		Client:          fakeClient,
-		Log:             logr.Discard(),
-		RuleStore:       ruleStore,
-		resourceCatalog: newCommonTestCatalog(t),
-		discoveryClient: commonTestDiscoveryClient(),
-	}
-
-	// Add a rule that references a non-existent GVR
-	watchRule := &configv1alpha1.WatchRule{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-missing-gvr",
-			Namespace: "default",
-		},
-		Spec: configv1alpha1.WatchRuleSpec{
-			TargetRef: configv1alpha1.LocalTargetReference{
-				Name: "test-dest",
-			},
-			Rules: []configv1alpha1.ResourceRule{
-				{
-					APIGroups:   []string{"custom.example.com"},
-					APIVersions: []string{"v1alpha1"},
-					Resources:   []string{"customresources"},
-				},
-			},
-		},
-	}
-
-	ruleStore.AddOrUpdateWatchRule(
-		*watchRule,
-		"test-dest", "default",
-		"test-repo", "default",
-		"main",
-		"test-folder",
-	)
-
-	// Compute requested GVRs. Catalog-backed resolution no longer requests a
-	// concrete informer for a resource that trusted discovery does not serve.
-	requestedGVRs := manager.ComputeRequestedGVRs()
-	if len(requestedGVRs) != 0 {
-		t.Fatalf("expected no requested GVRs for not-served resource, got %d", len(requestedGVRs))
-	}
-
-	t.Logf("Requested GVRs: %d", len(requestedGVRs))
-	for _, gvr := range requestedGVRs {
-		t.Logf("  - %s/%s %s", gvr.Group, gvr.Version, gvr.Resource)
-	}
-
-	// Filter discoverable - should return empty since CRD doesn't exist
-	discoverableGVRs := manager.FilterDiscoverableGVRs(ctx, requestedGVRs)
-
-	t.Logf("Discoverable GVRs: %d", len(discoverableGVRs))
-
-	// In fake client without full API discovery, this will be 0
-	// The real test is in e2e where actual discovery happens
-	if len(discoverableGVRs) > len(requestedGVRs) {
-		t.Errorf("Discoverable GVRs (%d) should not exceed requested GVRs (%d)",
-			len(discoverableGVRs), len(requestedGVRs))
-	}
-
-	t.Log("✅ Unavailable GVR tracking test completed")
-}
-
-// TestReconcileAfterRuleCreation tests that ReconcileForRuleChange is called immediately
-// after a rule is created, not just on periodic reconciliation.
-func TestReconcileAfterRuleCreation(t *testing.T) {
-	scheme := runtime.NewScheme()
-	_ = clientgoscheme.AddToScheme(scheme)
-	_ = configv1alpha1.AddToScheme(scheme)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	ruleStore := rulestore.NewStore()
-
-	// Track reconciliation calls
-	reconciledChan := make(chan struct{}, 1)
-	mockManager := &mockWatchManager{
-		reconcileFn: func(_ context.Context) error {
-			select {
-			case reconciledChan <- struct{}{}:
-			default:
-			}
-			return nil
-		},
-	}
-
-	// Simulate what WatchRule controller does
-	t.Log("Simulating WatchRule creation")
-
-	// 1. Add rule to store
-	watchRule := &configv1alpha1.WatchRule{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "immediate-test",
-			Namespace: "default",
-		},
-		Spec: configv1alpha1.WatchRuleSpec{
-			TargetRef: configv1alpha1.LocalTargetReference{
-				Name: "test-dest",
-			},
-			Rules: []configv1alpha1.ResourceRule{
-				{
-					APIGroups: []string{""},
-					Resources: []string{"configmaps"},
-				},
-			},
-		},
-	}
-
-	ruleStore.AddOrUpdateWatchRule(
-		*watchRule,
-		"test-dest", "default",
-		"test-repo", "default",
-		"main",
-		"test-folder",
-	)
-
-	// 2. Trigger reconciliation (what the controller does)
-	if err := mockManager.ReconcileForRuleChange(ctx); err != nil {
-		t.Fatalf("ReconcileForRuleChange failed: %v", err)
-	}
-
-	// 3. Verify reconciliation was called immediately
-	select {
-	case <-reconciledChan:
-		t.Log("✅ Reconciliation triggered immediately after rule creation")
-	case <-time.After(100 * time.Millisecond):
-		t.Error("❌ Reconciliation was not triggered immediately after rule creation")
-	}
-}
-
-// mockWatchManager is a test double for testing reconciliation calls.
-type mockWatchManager struct {
-	reconcileFn func(context.Context) error
-}
-
-func (m *mockWatchManager) ReconcileForRuleChange(ctx context.Context) error {
-	if m.reconcileFn != nil {
-		return m.reconcileFn(ctx)
-	}
-	return nil
-}
-
-func (m *mockWatchManager) GetClusterStateForGitDest(
-	_ context.Context,
-	_ types.NamespacedName,
-) ([]interface{}, error) {
-	return nil, nil
-}
-
-// TestPeriodicRediscovery tests that CRDs become available through periodic reconciliation
-// even if they weren't available initially.
-func TestPeriodicRediscovery(t *testing.T) {
-	t.Log("This test verifies the periodic reconciliation behavior")
-	t.Log("In production, periodic reconciliation runs every 30 seconds")
-	t.Log("The fix will add immediate reconciliation on CRD events")
-
-	// This is more of a documentation test - the actual behavior is tested in e2e
-	// where we can verify:
-	// 1. CRD installed → WatchRule created → immediate discovery ✅
-	// 2. WatchRule created → CRD installed → periodic discovery catches it ✅
-
-	t.Log("✅ Periodic rediscovery documented - see e2e tests for actual verification")
-}
