+++ internal/watch/manager_catalog.go
@@ -28,10 +28,25 @@ import (
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/dynamic/dynamicinformer"
+	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/cache"
+	ctrl "sigs.k8s.io/controller-runtime"
 
 	configv1alpha1 "github.com/ConfigButler/gitops-reverser/api/v1alpha1"
 )
+
+// restConfig acquires the controller runtime REST config.
+// Returns nil if no config is available (e.g., in unit tests without a cluster).
+func (m *Manager) restConfig() *rest.Config {
+	// ctrl.GetConfig reads KUBECONFIG or in-cluster config. In tests/e2e this is
+	// set up by the test harness/Kind. In unit tests without a cluster it returns
+	// an error, which callers handle gracefully.
+	cfg, err := ctrl.GetConfig()
+	if err != nil {
+		return nil
+	}
+	return cfg
+}
 
 func crdTriggerGVR() schema.GroupVersionResource {
 	return schema.GroupVersionResource{
